refactor(httputil): extract duplicate parseHeaders into shared package
The parseHeaders function was identically duplicated in net/net.go and
notifications/webhook.go. Extract it into httputil.ParseHeaders to
eliminate duplication. New httputil package avoids import cycles with
the existing utils package.
