+++ httputil/headers.go
@@ -1,0 +1,17 @@
+package httputil
+
+import "strings"
+
+// ParseHeaders converts a slice of "Key: Value" headers into a map.
+func ParseHeaders(headers []string) map[string]string {
+	headerMap := make(map[string]string, len(headers))
+	for _, header := range headers {
+		parts := strings.SplitN(header, ":", 2)
+		if len(parts) == 2 {
+			key := strings.TrimSpace(parts[0])
+			value := strings.TrimSpace(parts[1])
+			headerMap[key] = value
+		}
+	}
+	return headerMap
+}
