+++ net/net_test.go
@@ -156,51 +156,6 @@ func TestAutoDetectProtocol(t *testing.T) {
 	}
 }
 
-func TestParseHeaders(t *testing.T) {
-	tests := []struct {
-		name     string
-		headers  []string
-		expected map[string]string
-	}{
-		{
-			name:     "valid headers",
-			headers:  []string{"Content-Type: application/json", "Authorization: Bearer token"},
-			expected: map[string]string{"Content-Type": "application/json", "Authorization": "Bearer token"},
-		},
-		{
-			name:     "headers with spaces",
-			headers:  []string{" Content-Type : application/json "},
-			expected: map[string]string{"Content-Type": "application/json"},
-		},
-		{
-			name:     "malformed header skipped",
-			headers:  []string{"valid: header", "invalid-header", "another: valid"},
-			expected: map[string]string{"valid": "header", "another": "valid"},
-		},
-		{
-			name:     "empty headers",
-			headers:  []string{},
-			expected: map[string]string{},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := parseHeaders(tt.headers)
-			if len(got) != len(tt.expected) {
-				t.Errorf("parseHeaders() returned %d headers, want %d", len(got), len(tt.expected))
-			}
-			for key, expectedValue := range tt.expected {
-				if gotValue, exists := got[key]; !exists {
-					t.Errorf("parseHeaders() missing key %q", key)
-				} else if gotValue != expectedValue {
-					t.Errorf("parseHeaders() key %q = %q, want %q", key, gotValue, expectedValue)
-				}
-			}
-		})
-	}
-}
-
 func TestFormatURL(t *testing.T) {
 	tests := []struct {
 		name  string
